Refactors ceph_cos_agent lib as a charm-lib
