+++ ceph-mon/unit_tests/test_ceph_metrics.py
@@ -12,7 +12,6 @@ import unittest
 
 from ops.testing import Harness
 
-import ceph_metrics  # noqa: avoid circ. import
 import charm
 
 
@@ -155,17 +154,21 @@ class TestCephCOSAgentProvider(CephMetricsTestBase):
             "cos-agent",
         )
 
-    @patch("ceph_metrics.mgr_config_set_rbd_stats_pools", lambda: None)
-    @patch("ceph_metrics.ceph_utils.is_bootstrapped", return_value=True)
-    @patch("ceph_metrics.ceph_utils.is_mgr_module_enabled", return_value=False)
-    @patch("ceph_metrics.ceph_utils.mgr_enable_module")
-    @patch("ceph_metrics.ceph_utils.mgr_disable_module")
+    @patch("charms.ceph_mon.v0.ceph_cos_agent."
+           "CephCOSAgentProvider.mgr_config_set_rbd_stats_pools")
+    @patch("charms.ceph_mon.v0.ceph_cos_agent."
+           "ceph_utils.is_bootstrapped", return_value=True)
+    @patch("charms.ceph_mon.v0.ceph_cos_agent."
+           "ceph_utils.is_mgr_module_enabled", return_value=False)
+    @patch("charms.ceph_mon.v0.ceph_cos_agent.ceph_utils.mgr_enable_module")
+    @patch("charms.ceph_mon.v0.ceph_cos_agent.ceph_utils.mgr_disable_module")
     def test_add_remove_rel(
         self,
         mgr_disable_module,
         mgr_enable_module,
         _is_mgr_module_enable,
         _is_bootstrapped,
+        mgr_config_set_rbd_stats_pools,
     ):
         rel_id = self.harness.add_relation("cos-agent", "grafana-agent")
         self.harness.add_relation_unit(rel_id, "grafana-agent/0")
@@ -181,6 +184,7 @@ class TestCephCOSAgentProvider(CephMetricsTestBase):
         self.assertTrue("metrics_alert_rules" in data)
         self.assertTrue("groups" in data["metrics_alert_rules"])
         mgr_enable_module.assert_called_once()
+        mgr_config_set_rbd_stats_pools.assert_called_once()
 
         self.harness.remove_relation(rel_id)
         mgr_disable_module.assert_called_once()
